pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Aleks-Ti <mti61ro@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "1.0.0"
bitarray = "2.8.0"
aiogram = "^3.4.1"
sqlalchemy = "^2.0.28"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
